Change from usize to u64
On [random-access-storage](https://github.com/datrs/random-access-storage)
issue
https://github.com/datrs/random-access-storage/issues/6 has changed all
types from `usize` into `u64` to be able to handle more than 4gbs on
32bits systems.

> usize is 32 bits on a 32 bit system, so the storage would be limited
to 4GB on such a system.

When changing `random-access-storage` on `hypercore` (tracking:
https://github.com/datrs/hypercore/pull/100) one of the things I've
noticed is that it wold be easier to also make `merkle-tree-stream` use
`u64` to integrate with `random-access-storage`. Very likely, this also
means that we would need to bump to use more than 32Gb storages on
`merkle-tree-stream`.

I've simply changed all declarations of usize into u64. All tests are
passing.

Needs:
---

- [ ] https://github.com/datrs/flat-tree/pull/44

+++ src/default_node.rs
@@ -5,15 +5,15 @@ use std::ops::{Deref, DerefMut};
 #[derive(Debug, Eq, Ord, PartialEq, PartialOrd, Clone)]
 pub struct DefaultNode {
   /// Reference to this node's parent node.
-  pub parent: usize,
+  pub parent: u64,
   /// Data if it's a leaf node, nothing if it's a parent node.
   pub data: Option<Vec<u8>>,
   /// Hash of the data
   pub hash: Vec<u8>,
   /// Total size of all its child nodes combined.
-  pub length: usize,
+  pub length: u64,
   /// Offset into the flat-tree data structure.
-  pub index: usize,
+  pub index: u64,
 }
 
 impl DefaultNode {
@@ -39,7 +39,7 @@ impl Node for DefaultNode {
     &self.hash
   }
 
-  fn len(&self) -> usize {
+  fn len(&self) -> u64 {
     self.length
   }
 
@@ -47,11 +47,11 @@ impl Node for DefaultNode {
     self.length == 0
   }
 
-  fn index(&self) -> usize {
+  fn index(&self) -> u64 {
     self.index
   }
 
-  fn parent(&self) -> usize {
+  fn parent(&self) -> u64 {
     self.parent
   }
 }
